Support CSV data type for global quote and search Add tests for global quotes and search Improve media type detection Retain HttpResponse in Response

+++ src/main/kotlin/model/stock/SearchMatch.kt
@@ -13,6 +13,6 @@ data class SearchMatch(
         val marketClose: LocalTime,
         val timeZone: TimeZone,
         val currency: Currency,
-        val matchScore: Float,
+        val matchScore: Double,
         ) {
 }
